refactor:회원가입타이머컴포넌트화
